+++ unifi_talk_alarm/app/talk_alarm/__init__.py
@@ -1,3 +1,3 @@
 """UniFi Talk Alarm add-on service."""
 
-__version__ = "0.2.1"
+__version__ = "0.2.2"